requirements.txt
anyio==3.6.2
asgiref==3.6.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
docx2txt==0.8
enchant==0.0.1
filelock==3.12.0
fsspec==2023.4.0
googlesearch-python==1.2.3
grpcio==1.54.0
grpcio-tools==1.54.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.14.1
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pdfkit==1.0.0
pdfminer.six==20221105
pdfplumber==0.10.2
Pillow==9.5.0
protobuf==4.22.3
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pypdfium2==4.20.0
python3-openid==3.2.0
PyYAML==6.0
qdrant-client==1.1.6
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.3
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
